feature: unpack network bytes into usable format
use the dns header struct to parse the dns packets L7 header data.
add loop on bytes to parse the bytes into a human readable string.

- properly calc payload length.
after the HEADER the packet contains the QNAME sections. This will
contain the relevant data for making the decision whether we want to
resolce a hostname or not. There are two u16 values at the end of the
packet which should not be parsed into an utf-8 string.

- dns QNAME bytes: will have a byte sequence with this format
<u8: length of next segmen><ASCII CODE FOR first char of seqment>,...

e.g.:
`aaaa.bcd.cc`
is represented by:

[<4>, <97>,<97>,<97>,<97>, <3>, <98>,<99>,<100>, <2>,<99>,<99>]