switch ingress to debug log incoming packet data
the xpd part is currently unused. But it could be useful down the road.
currently the blocked requests will get dropped. but depending on the
application that sends those requests, these will be retried until a
timeout / infinity. it would be nicer to figure out the way to redirect
the packet to a fake host and use the xdp ingress to refuse the
connection or something like that in hopes of no retries from the
client.