feat: implement redirect
the eBPF Program will now redirect the packet to localhost. Unfortunatly this will
still time out, even if we use 0.0.0.0 as the target.

Though we can use the XDP context to wait for that redirected packet and instantly
drop it. Basically just need to find a free port on the system so we don't accidentially
drop other stuff.