feat[wip]]: rely on dns caching to limit amount of requests
the client that fires the query should be caching the response according
to the TTL in the DNS response. Maybe redirection of the packets into
nirvana isn't the best way to go, since it was found in testing that the
requests for bad addr. are usually retried.

This adds parsing of the response received from upstream on the XPD
program to allow to rewrite the received packet with:

- a fake IP
- a max_val(u32) for TTL ( which hopefully will get rid of the retries
  from the client )